cv-analyzer/requirements/requirements.txt
# Web Framework
flask==2.2.3
flask-cors==3.0.10
gunicorn==20.1.0

# File processing
python-docx==0.8.11
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==9.5.0

# NLP and text analysis
spacy==3.5.2
nltk==3.8.1
pyresparser==1.0.6
scikit-learn==1.2.2
sentence-transformers==2.2.2

# Firebase
firebase-admin==6.1.0

# Utilities
python-dotenv==1.0.0
requests==2.28.2
numpy==1.24.2
pandas==2.0.0
